#164 Aplicando sugestoes de correcao
