+++ lib/controllers/profile/profile_controller.dart
@@ -1,10 +1,9 @@
 import 'package:caderneta_campo_digital/global/global.dart';
-import 'package:caderneta_campo_digital/models/models.dart';
+import 'package:caderneta_campo_digital/models/UserModel.dart';
 import 'package:flutter/material.dart';
 import '../../pages/login/login_page.dart';
 
 class ProfileController {
-
   /*ProfileService profileService = ProfileService();
   String errorText = "";
   bool loading = false;
@@ -36,9 +35,9 @@ class ProfileController {
     );
 
     Navigator.pushAndRemoveUntil(
-        context,
-        MaterialPageRoute(builder: (context) => LoginPage()),
-        (route) => route.isCurrent,);
+      context,
+      MaterialPageRoute(builder: (context) => LoginPage()),
+      (route) => route.isCurrent,
+    );
   }
-
 }
