+++ lib/models/UserModel.dart
@@ -44,42 +44,3 @@ class UserModel {
   factory UserModel.fromJson(String source) =>
       UserModel.fromMap(json.decode(source));
 }
-
-class AgrotoxicType {
-  final String id;
-  final String name;
-
-  AgrotoxicType(this.id, this.name);
-
-  static AgrotoxicType fromJson(Map<String, dynamic> data) => AgrotoxicType(
-        data["idTipoAgrotoxico"],
-        data["nome"],
-      );
-
-  @override
-  String toString() {
-    return name;
-  }
-}
-
-class Agrotoxic {
-  final String id;
-  final String name;
-  final AgrotoxicType type;
-
-  Agrotoxic(this.id, this.name, this.type);
-
-  static Agrotoxic fromJson(Map<String, dynamic> data) => Agrotoxic(
-        data["idAgrotoxico"],
-        data["nome"],
-        AgrotoxicType(
-          data["tipo"]["idTipoAgrotoxico"],
-          data["tipo"]["nome"],
-        ),
-      );
-
-  @override
-  String toString() {
-    return name;
-  }
-}
